Move OutputMetadata from client to types
