+++ client/src/api/block_builder/input_selection/manual.rs
@@ -8,7 +8,7 @@ use std::collections::HashSet;
 use crypto::keys::slip10::Chain;
 use iota_types::block::{
     address::Address,
-    output::{AliasId, Output},
+    output::{AliasId, Output, OutputMetadata},
     protocol::ProtocolParameters,
 };
 
@@ -19,7 +19,7 @@ use crate::{
         ClientBlockBuilder,
     },
     constants::HD_WALLET_TYPE,
-    secret::types::{InputSigningData, OutputMetadata},
+    secret::types::InputSigningData,
     Result,
 };
